Catalog: single sidebar; Buckets is a link to home; Bookmarks opens a right drawer
