+++ catalog/app/containers/Bookmarks/index.ts
@@ -1,2 +1,3 @@
 export * from './Provider'
 export { List } from './List'
+export { Drawer } from './Drawer'
